refactor: #58 사용자 최초 로그인 여부 데이터까지 한번에 조회

+++ RevUp-api/src/main/java/com/revup/answer/dto/response/AnswerDetailsResponse.java
@@ -23,7 +23,7 @@ public record AnswerDetailsResponse(
                 answer.getUser().getNickname(),
                 answer.getUser().getProfileImage(),
                 answer.getUser().getTotalAnswerCount(),
-                answer.getUser().getAdoptedRate(),
+                answer.getUser().getAdoptedRate().intValue(),
                 answer.getCreatedAt().toString(),
                 answer.getIsAccept().toBoolean(),
                 answer.getGoodCount(),
